requirements.txt
streamlit>=1.28.0
numpy>=1.24.3
opencv-python-headless>=4.8.0
Pillow>=10.0.1
torch>=2.0.1
transformers>=4.33.1
jiwer>=3.0.3
pyspellchecker>=0.8.0
python-dotenv>=1.0.0
craft-text-detector>=0.4.3
requests>=2.31.0
texteller
